New OpenBCI packet format, both 8 and 16 channels.
Including XYZ Accelerometer values.

Sync loss errors are logged to GLOBAL.syncloss++ counter so that
they show up on status bar at the bottom of the window.

If a resync is needed, search for END followed by START byte,
so we have a bit more robust restart.

'GAIN' calculation now done with a #define in case that changes.

Disable changing the 'resolution' field in the dialog box, since
Chris says that was only used long ago.

Send the 'Stop' byte when session is stopped.